Add files via upload
Simple bank account system that allows the user to create a new account and input their full name, ID number and gender. The system then generates an account number. It also allows the user to update their account information by depositing or withdrawing. Yet to be more secure because right now, it does not verify if account number is unique and it doesn't match with others.